==> 01_ETL_Output.ipynb <==
## Cargamos las librerias

import pandas as pd
import numpy as np
import json
import warnings
warnings.filterwarnings("ignore")

## Cargamos el archivo json y pasamos los datos a un dataframe

ruta_output= "Data2\output_steam_games.json"
## with open(ruta_output, 'r') as archivo:
##    datos_output = json.load(archivo)
# Este codigo no sirve dado que hay datos adicionales en el archivo JSON después de que se esperaba que finalizara el documento JSON.
# Por lo que se procede a leer linea por linea para extraer la informacion

with open(ruta_output, 'r') as archivo: #"r" indica que se abre en modo lectura y "with" garantiza que despues de su uso se cierre correctamente
    datos_output = [json.loads(line) for line in archivo]

# Transformamos a un dataframe
df_output = pd.DataFrame(datos_output)
df_output

df_output.columns

df_output.info()

Observamos que todos nuestros tipos de datos son objetos

df_output["tags"][120000]

df_output["early_access"].unique()

df_output["price"].unique()

df_output["specs"][120000]

df_output["genres"][120000]

## Verficamos datos nulos

df_output.isnull().sum()

Eliminamos las filas que solo contienen valores nulos es decir que no aportan nada

df_output = df_output.dropna(how="all").reset_index(drop=True)

df_output.shape

Volvemos a verificar que variables tienen valores nulos y cuanto representan

cantidad_nulos= df_output.isnull().sum()
print(cantidad_nulos)

Aqui eliminamos los valores nulos que se encuentran en la variable id

df_output.dropna(subset=['id'], inplace=True)

cantidad_filas = len(df_output)
porcentaje_nulos_columnas= (cantidad_nulos/cantidad_filas)*100
porcentaje_nonulos_columnas = 100 - porcentaje_nulos_columnas
resumen_nulos = pd.DataFrame({"Porcentaje de valores nulos":porcentaje_nulos_columnas, 
                              "Porcentaje de valores no nulos":porcentaje_nonulos_columnas})
resumen_nulos

## Filas duplicadas

En este caso la columna en la que se quiere verificar si tiene duplicados es la columna .
"id" dado que es una variable identificatoria de usuario

df_output[df_output["id"].duplicated()]

Eliminamos los datos duplicados pertenecientes a id

df_output = df_output.drop_duplicates(subset='id', keep='first')

df_output.isnull().sum()

## Variable Price

En este caso la columna como evidenciamos antes tiene dos tipos de datos que son str y float, para poder trabajar con ella necesitamos que los valores queden como float analizando nos damos cuenta que los valores que estan en str indican que el juego fue gratis por lo que le vamos a asignar un precio de cero

df_output["price_number"] = pd.to_numeric(df_output["price"], errors="coerce")
df_output["price_number"] = pd.to_numeric(df_output["price_number"], errors="coerce")
df_output["price_number"] = df_output["price_number"].fillna(0)

df_output["price_number"].dtype

## Variable Release date

En este caso lo que a nosotros nos va a importar es el año de lanzamiento por lo que vamos a crear una nueva columna para este fin

df_output["release_date"].isnull().sum()

df_output["release_date"].unique()

invalid_dates = df_output["release_date"][~df_output["release_date"].str.match(r"\d{4}-\d{2}-\d{2}", na=False, case=False, flags=0)]
df_invalid_dates = pd.DataFrame({"invalid_dates": invalid_dates})
df_invalid_dates

En este caso cambiamos el formato de la columna release_date para luego crear la columna release_year que es la variable que nos importa para el proximo analisis

df_output["release_date"] = pd.to_datetime(df_output["release_date"], errors="coerce", format='%Y-%m-%d', exact=False)
mean_release_date = df_output["release_date"].mean()
df_output["release_date_filled"] = df_output["release_date"].fillna(mean_release_date)
df_output["release_year"] = df_output["release_date_filled"].dt.year
df_output

df_output["release_year"].dtype

df_output["release_year"].unique()

df_output["release_year"].isnull().sum()

### Eliminamos las columnas que no seran necesarias para nuestro analisis y posterior modelado para optimizar el dataframe

df_output.drop(["release_date", "price","url","publisher","reviews_url","specs","early_access","release_date_filled"], axis=1, inplace=True)
df_output

## Variable Genres

df_output["genres"].isnull().sum()

Vamos a imputar los valores nulos de la variable genres con los valores pertenecientes a la variable tags pero los cuales si pertenecen a un genero

generos = []
for genres_list in df_output["genres"].dropna():
    for genre in genres_list:
        if genre not in generos:
            generos.append(genre)
generos

df_output["tags"]=df_output["tags"].dropna().apply(lambda lista: [g for g in lista if g in generos])

df_output["genres"].fillna(df_output["tags"], inplace=True)

df_output["genres"].isnull().sum()

Aunque imputamos los registros podemos verificar que siguen quedando registros nulos que se trataran mas adelante

Procedemos a eliminar la columna tags dado que ya no es necesaria 

df_output = df_output.drop("tags", axis=1)
df_output

df_output["genres"].head()

Como podemos observar la variable contiene una lista de generos en cada registro por lo que nos encontramos con datos anidados que es necesario desanidarlos, en este caso para realizar el cometido se utilizara explode para desanidar 

df_output["main_genre"] = df_output["genres"].apply(lambda x: x[0] if isinstance(x, list) and x else None)
df_unique_genre = df_output.drop_duplicates(["app_name", "title", "id", "developer", "price_number", "release_year"])
df_unique_genre = df_unique_genre.drop("genres", axis=1)
df_unique_genre

df_unique_genre["main_genre"].isnull().sum()

df_unique_genre["main_genre"] = df_unique_genre["main_genre"].fillna("Sin Dato")
df_unique_genre["main_genre"].isnull().sum()

## Variables App_name, title y developer

columnas_a_llenar = ["app_name", "title"]
df_unique_genre[columnas_a_llenar] = df_unique_genre[columnas_a_llenar].fillna("Sin datos")


son_iguales = (df_unique_genre["app_name"] == df_unique_genre["title"]).all()

if son_iguales:
    print("Las columnas 'app_name' y 'title' son iguales en todos los registros.")
else:
    print("Las columnas 'app_name' y 'title' no son iguales en todos los registros.")


df_unique_genre["son_iguales"] = df_unique_genre["app_name"] == df_unique_genre["title"]
filas_no_iguales = df_unique_genre[df_unique_genre["son_iguales"] == False]
filas_no_iguales

En el caso de los Developers es una variable importante para el analisis posterior por lo que los valores nulos que no aportan informacion se proceden a eliminar

df_unique_genre.dropna(subset=["developer"], inplace= True)

df_unique_genre.isnull().sum()

Como pudimos evidenciar las columnas app_name y title solo difieren que en la columna title los registros son diferentes es porque tienen las letras amp por lo que solo nos vamos a quedar con la columna app_name

df_unique_genre.drop(columns=["title"],axis=1,inplace=True)

df_unique_genre

df_unique_genre.drop(["son_iguales"], axis=1, inplace=True)

df_unique_genre.head(2)

df_unique_genre.isnull().sum()

## Cargamos el dataset limpio

archivo_limpio = 'data2/output_steam_games_limpio9.csv'
df_unique_genre.to_csv(archivo_limpio, index=False, encoding='utf-8')
print(f'Se guardó el archivo {archivo_limpio}')
