==> 01_ETL_user_items.ipynb <==
## Cargamos las librerias necesarias

import pandas as pd
from pandas import json_normalize
import numpy as np
import json
import warnings
warnings.filterwarnings("ignore")
import ast

## Cargamos el archivo json y pasamos los datos a un dataframe

ruta_items = "data2/australian_users_items.json"
filas_items = []
with open(ruta_items, encoding="utf-8") as f:
    for line in f.readlines():
        filas_items.append(ast.literal_eval(line))

# Transformamos a un dataframe
df_items = pd.DataFrame(filas_items)
df_items

df_items.columns

### Se observa los tipos de objetos que presenta el dataframe y cantidad de datos nulos

df_items.info()

### Se procede a observar variable por variable para entender los registros guardados

Esta variable representa el identificador unico del usuario

df_items["user_id"][25]

"items_count" representa la cantidad de juegos consumidos por cada usuario

df_items["items_count"][25]

df_items["steam_id"][25]

"user_url" indica la url que te dirige al perfil del usuario

df_items["user_url"][25]

"items" representa los juegos que cada usuario consumio junto a las caracteristicas de los mismos

df_items["items"][25]

Como se observo en la linea de codigo anterior la variable items contiene varios diccionarios dentro de un mismo registro por lo que se procede a desanidar la misma para su posterior analisis

df_items= df_items.explode("items").reset_index(drop=True)
df_items

df_items_normalizado = pd.json_normalize(df_items["items"])
df_items_normalizado

df_items_final = pd.concat([df_items,df_items_normalizado], axis=1)
df_items_final

Eliminamos la columna items dado que ya no es necesaria

df_items_final = df_items_final.drop(columns=["items"])
df_items_final

### Se procede a eliminar columnas que no son relevantes para nuestro proposito como "user_url" y "playtime_2weeks"

df_items_final = df_items_final.drop(columns=["user_url","playtime_2weeks","steam_id"])
df_items_final

df_items_final.info()

### Corroboramos sobre la variable "item_name" que tengamos registros que se puedan usar para el modelado

df_items_final["item_name"].unique()

patron_no_deseado = r'[^a-zA-Z0-9\s]'  
df_filtrado = df_items_final[df_items_final["item_name"].str.contains(patron_no_deseado) == False]
df_filtrado

### Se verifica si tenemos registros nulos

df_filtrado.isnull().sum()

### Se verifica si tenemos registros duplicados

duplicados = df_filtrado.duplicated().sum()
duplicados

duplicados_mask = df_filtrado.duplicated(keep=False)
filas_duplicadas = df_filtrado[duplicados_mask]
filas_duplicadas

En este caso verificamos que las filas que aparecen como duplicadas corresponden al mismo usuario por lo que las variables: user_id, items_count, steam_id son las mismas pero varia el juego al cual se esta refiriendo por lo que se decide conservarlas

### Carga del conjunto de datos

archivo_limpio = "data2/user_items_limpio5.csv"
df_filtrado.to_csv(archivo_limpio, index=False, encoding="utf-8")
print(f"Se guardó el archivo {archivo_limpio}")
