==> 01_ETL_user_reviews.ipynb <==
## Cargamos las librerias

import pandas as pd
from pandas import json_normalize
import numpy as np
import json
import warnings
warnings.filterwarnings("ignore")
import ast
import re
from textblob import TextBlob

## Cargamos el archivo json y pasamos los datos a un dataframe

ruta_reviews = "data2/australian_user_reviews.json"
filas_review = []
with open(ruta_reviews, encoding='utf-8') as f:
    for line in f.readlines():
        filas_review.append(ast.literal_eval(line))

# Transformamos a un dataframe
df_reviews = pd.DataFrame(filas_review)
df_reviews

df_reviews.info()

df_reviews.isnull().sum()

Podemos verificar que todos nuestros datos son de tipo objeto y que no tenemos ninguno que sea nulo

Verificamos si hay filas duplicadas en funcion del user_id y si son validas

duplicados = df_reviews[df_reviews.duplicated("user_id", keep=False)]
duplicados

Podemos observar que tenemos 623 registros duplicados en cuanto a user_id pero puede ser que un mismo usuario haya realizado mas de un comentario por lo que se procede a analizar

user_id = "76561198156664158"
user_rows = duplicados.loc[duplicados["user_id"] == user_id, "reviews"]
for review_list in user_rows:
    for review in review_list:
        print(review)
    print("-" * 40)

Aqui se tomo un usuario para verificar si era el mismo comentario repetido y estabamos en lo correcto por lo que se procede a eliminar

df_reviews = df_reviews.drop_duplicates(subset="user_id", keep=False)
df_reviews

En la variable reviews se observo que muchos de los registros se encuentran de forma anidada se analiza la estructura


df_reviews["reviews"][0]

Corresponde desanidar la columna para luego poder proceder con el analisis. En este caso procedemos con la funcion "explode" para poder convertir cada elemento de la lista en una fila separada, replicando el índice y las demás columnas según sea necesario.

df_reviews_expanded = df_reviews.explode("reviews")
df_reviews_normalized = json_normalize(df_reviews_expanded["reviews"])
df_reviews_expanded.reset_index(drop=True, inplace=True)
df_reviews_normalized.reset_index(drop=True, inplace=True)
df_reviews_final = pd.concat([df_reviews_expanded,df_reviews_normalized], axis=1)
df_reviews_final = df_reviews_final.drop(columns='reviews')
df_reviews_final.head(2)

df_reviews_final.shape

df_reviews_final.info()

En el  caso de las nuevas columnas generadas observamos que no tenemos valores nulos pero que si hay registros que no tienen informacion por lo que se verifica

conteo_valores_vacios = df_reviews_final.apply(lambda col: col[col == ""].count())
conteo_valores_vacios

Solo tres columnas contienen valores vacios en el caso de la columna "funny" y "last_edited" no son necesarias para nuestro analisis por lo que se procede a eliminarlas y en cuanto a "reviews" se concluye que varios usuarios no dieron un comentario por lo que se coloca None. Tambien se observa que algunos registros de review contiene caracteres que pueden afectar el modelado posterior por lo que se corrigen

df_reviews_final = df_reviews_final.drop(columns=["funny","last_edited"])
df_reviews_final

df_reviews_final = df_reviews_final.drop(columns=["user_url"])
df_reviews_final

df_reviews_final.replace("",None, inplace=True)
df_reviews_final.head()

df_reviews_final.info()

review_nulos = df_reviews_final["review"].isnull().sum()
print(review_nulos)

nulos_largo = len(df_reviews_final["review"])
porcentaje_nulos_review = (review_nulos/nulos_largo)*100
porcentaje_nulos_review

df_reviews_final=df_reviews_final.dropna(subset=["review"])
df_reviews_final = df_reviews_final.reset_index(drop=True)
df_reviews_final.info()

### Verificamos los registros que tenemos en review

reviews_col = df_reviews_final["review"].head(60)
reviews_col

def clean_and_show(text):
    cleaned_text = re.sub(r"[^a-zA-Z\s]", "", text)
    if cleaned_text != text:
        print(f"Texto original: {text}")
    return cleaned_text

df_reviews_final["review"] = df_reviews_final["review"].apply(clean_and_show)
# Esta funcion devuelve las filas que contienen caracteres especiales

### Verificamos que la funcion haya cumplido con el proposito

reviews_col = df_reviews_final["review"].head(60)
reviews_col

### De la variable "posted" solo necesitamos el año en el que se publico la review por lo que se procede a extraer para luego eliminarla

df_reviews_final["year_posted"] = df_reviews_final["posted"].str.extract(r'(\d{4})') 
df_reviews_final

df_reviews_final.drop("posted" , axis = 1, inplace = True)
df_reviews_final

nulos_year_posted = df_reviews_final.isnull().sum()
nulos_year_posted

df_reviews_final["year_posted"].value_counts(normalize=True)

Para no tener valores nulos vamos imputar aquellos registros nulos con la mediana dado que esta es menos sensible a valores extremos

df_reviews_final["year_posted"].fillna(2013, inplace=True)

df_reviews_final["year_posted"].isnull().sum()

Terminamos verificando si quedan valores nulos en alguna variable por revisar

df_reviews_final.isnull().sum()

### Variable "sentiment_analysis"

Crearemos la columna 'sentiment_analysis' aplicando análisis de sentimiento con NLP con la siguiente escala: valor '0' si es malo, '1' si es neutral y '2' si es positivo. Esta nueva columna reemplazara la de user_reviews.review para facilitar el trabajo de los modelos de machine learning y el análisis de datos.

def analyze_sentiment(text):
    if not isinstance(text, str):
        return 1 
    if review is None:
        return 1
    blob = TextBlob(text)
    sentiment_polarity = blob.sentiment.polarity
    if sentiment_polarity < -0.1:
        return 0
    elif sentiment_polarity > 0.1:
        return 2
    else:
        return 1
    
df_reviews_final["sentiment_analysis"] = df_reviews_final["review"].apply(analyze_sentiment)

Analizamos el funcionamiento de la nueva funcion creada

for index, row in df_reviews_final.head(5).iterrows():
    review_text = row["review"]
    sentiment = row["sentiment_analysis"]
    print(f"Review {index}:")
    print(f"Texto del Review: {review_text}")
    print(f"Análisis de Sentimiento (0-Negativo, 1-Neutral, 2-Positivo): {sentiment}")
    print("\n")

df_reviews_final.drop(columns=["review"], inplace=True)
df_reviews_final.head()

df_reviews_final.isnull().sum()

### Variable "helpful"

total_filas = df_reviews_final.shape[0]
no_reseña_helpful= df_reviews_final[df_reviews_final["helpful"]=="No ratings yet"]
num_rows_no_ratings = no_reseña_helpful.shape[0]
print(f"La cantidad de filas sin ratings es: {num_rows_no_ratings}")
porcentaje_no_ratings = (num_rows_no_ratings / total_filas) * 100
print(f"El porcentaje de filas sin ratings es: {porcentaje_no_ratings}")

Dado que la columna presenta la mitad de los registros sin ningun rating y ademas no es relevante para el posterior analisis y modelado, se procede a eliminarla

df_reviews_final.drop(columns=["helpful"], inplace=True)
df_reviews_final.head()

df_reviews_final.shape

### Exportamos el CSV limpio 

archivo_limpio = "data2/user_reviews_limpio7.csv"
df_reviews_final.to_csv(archivo_limpio, index=False, encoding="utf-8")
print(f"Se guardó el archivo {archivo_limpio}")
