==> 03_DataFrames_Auxiliares.ipynb <==
## Creacion de DataFrames auxiliares

En este notebook vamos a crear los data frames auxiliares que tendremos que usar para las funciones que vamos a desarrollar 

### Cargamos las librerias necesarias

import pandas as pd
import pyarrow as pa
import pyarrow.parquet as pq
import warnings
warnings.filterwarnings("ignore")

### Cargamos los CSV necesarios para extraer la informacion necesaria

df_reviews = pd.read_csv("data2/user_reviews_limpio7.csv", encoding="utf-8")
df_output = pd.read_csv("data2/output_steam_games_limpio9.csv", encoding="utf-8")
df_items = pd.read_csv("data2/user_items_limpio5.csv", encoding="utf-8")

df_items.columns

df_reviews.columns

df_output.columns

df_items["playtime_hour"]= df_items["playtime_forever"]/60
df_items

### Funcion PlayTimeGenre
En este caso nos piden una funcion que devuelva el año con mas horas jugadas para un genero especificado, por lo que seran necesarias las columnas main_genre, id y Release_year de df_output y la columna playtime_forever e item_id de df_items

df_output2 = df_output[["main_genre", "release_year","id"]]
df_items2 = df_items[["playtime_hour", "item_id"]]
df_PlayTimeGenre_merge = pd.merge(df_items2, df_output2, left_on='item_id', right_on="id")
df_PlayTimeGenre_hour = df_PlayTimeGenre_merge.groupby(["main_genre", "release_year"])["playtime_hour"].sum().reset_index()
df_PlayTimeGenre_hour

Guardamos el nuevo dataframe como un CSV

archivo_limpio = "data2/df_PlayTimeGenre_hour_final1.csv"
df_PlayTimeGenre_hour.to_csv(archivo_limpio, index=False, encoding="utf-8")
print(f"Se guardó el archivo {archivo_limpio}") 

Guardamos el nuevo dataframe como archivo PARQUET

archivo_limpio2 = "Datasets/df_PlayTimeGenre_hour_final1.parquet"
df_PlayTimeGenre_hour.to_parquet(archivo_limpio2, index=False)
print(f"Se guardó el archivo {archivo_limpio2}")

### Funcion UserForGenre
En este caso nos piden una funcion que devuelva el usuario que acumula más horas jugadas para el género dado y una lista de la acumulación de horas jugadas por año, por lo que seran necesarias las columnas "item_id","user_id" y "year_posted" de df_reviews, como tambien las columnas "user_id", "item_id", "playtime_hour" de df_items y "id" e "main_genre" de df_output

Filtramos las columnas de df_reviews y de df_items necesarias 

data_reviews_subset = df_reviews[["item_id","user_id"]]

data_items_subset = df_items[["user_id","item_id","playtime_hour"]]

Unimos las columnas que fueron previamente filtradas en un solo dataframe

merged_data = data_items_subset.merge(data_reviews_subset, on=["user_id", "item_id"], how="inner")

Al dataframe ya unido previamente le unimos las columnas necesarias de df_output

final_merged_data = merged_data.merge(df_output[["id","main_genre","release_year"]], left_on="item_id", right_on="id", how="inner")

Agrupamos apartir de "user_id", "main_genre" y "year_posted" y despues sumamos las horas de juegos acumuladas por estos grupos

df_UsersForGenre = final_merged_data.groupby(["user_id","main_genre","release_year"])["playtime_hour"].sum().reset_index()

df_UsersForGenre

data_reviews_subset2 = df_reviews[["item_id","user_id","year_posted"]]
data_items_subset2 = df_items[["user_id","item_id","playtime_hour"]]
merged_data2 = data_items_subset2.merge(data_reviews_subset2, on=["user_id", "item_id"], how="inner")
final_merged_data2 = merged_data2.merge(df_output[["id","main_genre"]], left_on="item_id", right_on="id", how="inner")
df_UsersForGenre2 = final_merged_data2.groupby(["user_id","main_genre","year_posted"])["playtime_hour"].sum().reset_index()
df_UsersForGenre2

archivo_limpio001 = "data2/df_UsersForGenre3_final.csv"
df_UsersForGenre2.to_csv(archivo_limpio001, index=False, encoding="utf-8")
print(f"Se guardó el archivo {archivo_limpio001}") 

Guardamos el nuevo dataframe como un CSV

archivo_limpio00 = "data2/df_UsersForGenre2_final.csv"
df_UsersForGenre.to_csv(archivo_limpio00, index=False, encoding="utf-8")
print(f"Se guardó el archivo {archivo_limpio00}") 

Guardamos el nuevo dataframe como archivo PARQUET

archivo_limpio000 = "Datasets/df_UsersForGenre2_final.parquet"
df_UsersForGenre.to_parquet(archivo_limpio000, index=False)
print(f"Se guardó el archivo {archivo_limpio000}")

### Funcion UsersRecommend
En este caso nos piden una funcion que devuelve el top 3 de juegos MÁS recomendados por usuarios para el año dado, por lo que seran necesarias la columna id y app_name de df_output y las columnas item_id, recommend y year_posted de df_reviews

df_output3_ = df_output[["id", "app_name"]].drop_duplicates()
df_reviews2_ = df_reviews[["item_id", "recommend", "year_posted","sentiment_analysis"]]
df_reviews_merged_ = pd.merge(df_reviews2_, df_output3_, left_on="item_id", right_on="id")
positive_reviews_ = df_reviews_merged_[(df_reviews_merged_["recommend"]== True) & (df_reviews_merged_["sentiment_analysis"] >=1)]
grouped_reviews_ = positive_reviews_.groupby(["year_posted", "item_id", "app_name"]).size().reset_index(name="recommendation_count")
df_UsersRecommend_ = (
    grouped_reviews_.sort_values(by=["year_posted", "recommendation_count"], ascending=[True, False])
                   .groupby("year_posted")
                   .head(3)
                   .reset_index(drop=True)
)
df_UsersRecommend_

Guardamos el nuevo dataframe como un CSV

archivo_limpio5 = "data2/df_UsersRecommend2_final.csv"
df_UsersRecommend_.to_csv(archivo_limpio5, index=False, encoding="utf-8")
print(f"Se guardó el archivo {archivo_limpio5}") 

Guardamos el nuevo dataframe como archivo PARQUET

archivo_limpio6 = "Datasets/df_UsersRecommend2_final.parquet"
df_UsersRecommend_.to_parquet(archivo_limpio6, index=False)
print(f"Se guardó el archivo {archivo_limpio6}")

### Funcion UsersWorstDeveloper
En este caso nos piden una funcion que devuelve el top 3 de desarrolladoras con juegos MENOS recomendados por usuarios para el año dado. (reviews.recommend = False y comentarios negativos). Se utilizaran las columnas "id" y "developer" del df_output, tambien las columnas "item_id", "recommend", "year_posted" y "sentiment_analysis"del df_reviews

df_output4 = df_output[["id", "developer"]].drop_duplicates()
df_reviews4 = df_reviews[["item_id", "recommend", "year_posted","sentiment_analysis"]]
df_reviews_merged4 = pd.merge(df_reviews4, df_output4, left_on="item_id", right_on="id")
negative_reviews4 = df_reviews_merged4[(df_reviews_merged4["recommend"]== False) & (df_reviews_merged4["sentiment_analysis"] == 0)]
grouped_reviews4 = negative_reviews4.groupby(["year_posted", "item_id", "developer"]).size().reset_index(name="recommendation_count")
df_UsersWorstDeveloper = (
    grouped_reviews4.sort_values(by=["year_posted", "recommendation_count"], ascending=[True, False])
                   .groupby("year_posted")
                   .head(3)
                   .reset_index(drop=True)
)
df_UsersWorstDeveloper

df_UsersWorstDeveloper_final = df_UsersWorstDeveloper.groupby(["year_posted","developer"], as_index=False)["recommendation_count"].sum()
df_UsersWorstDeveloper_final

Guardamos el nuevo dataframe como un CSV

archivo_limpio7 = "data2/df_UsersWorstDeveloper_final1.csv"
df_UsersWorstDeveloper_final.to_csv(archivo_limpio7, index=False, encoding="utf-8")
print(f"Se guardó el archivo {archivo_limpio7}") 

Guardamos el nuevo dataframe como archivo PARQUET

archivo_limpio8 = "Datasets/df_UserWorstDeveloper_final1.parquet"
df_UsersWorstDeveloper.to_parquet(archivo_limpio8, index=False)
print(f"Se guardó el archivo {archivo_limpio8}")

### Funcion Sentiment_analysis
En este caso nos piden una funcion que devuelve un diccionario con el nombre de la desarrolladora como llave y una lista con la cantidad total de registros de reseñas de usuarios que se encuentren categorizados con un análisis de sentimiento como valor. Se utilizan todas las columnas de df_reviews y las columnas "id","app_name" y "developer" de df_output para finalmente quedarnos con las columnas developer y sentiment_analysis y sobre estas columnas hacer un conteo del analisis de sentimiento

df_reviews0 = df_reviews[["user_id", "item_id","recommend","year_posted","sentiment_analysis"]]
df_reviews1 = pd.merge(df_reviews0, df_output[["id","app_name","developer"]], left_on="item_id", right_on="id")
df_reviews02 = df_reviews1[["user_id","app_name","developer","recommend","year_posted","sentiment_analysis"]]
df_Sentiment_analysis = df_reviews02.groupby(["developer","sentiment_analysis"]).size().reset_index(name="sentiment_analysis_count")
df_Sentiment_analysis

Guardamos el nuevo dataframe como un CSV

archivo_limpio10 = "data2/df_Sentiment_analysis_final.csv"
df_Sentiment_analysis.to_csv(archivo_limpio10, index=False, encoding="utf-8")
print(f"Se guardó el archivo {archivo_limpio10}") 

Guardamos el nuevo dataframe como archivo PARQUET

archivo_limpio11 = "Datasets/df_Sentiment_analysis_final.parquet"
df_Sentiment_analysis.to_parquet(archivo_limpio11, index=False)
print(f"Se guardó el archivo {archivo_limpio11}")
