==> 04_API_Funciones.ipynb <==
# Experimentación para la Elaboración de Funciones Cruciales para la API:
En esta fase, llevamos a cabo la creación y prueba de funciones esenciales para nuestra API, utilizando los datos disponibles. Este enfoque nos permite verificar la efectividad y eficiencia de las funciones en un entorno controlado. No obstante, al implementarlas finalmente, es posible que se requieran adaptaciones para cumplir con los requisitos específicos de FastAPI y Render.

import pandas as pd
import warnings
warnings.filterwarnings("ignore")

## Funcion PlayTimeGenre

df_PlayTimeGenre = pd.read_csv("data2/df_PlayTimeGenre_hour_final1.csv")
df_PlayTimeGenre.columns

def PlayTimeGenre( genero : str ):
    """
    Funcion que devuelve el año con mas horas jugadas para dicho género.
    """
    generos = df_PlayTimeGenre[df_PlayTimeGenre["main_genre"]== genero] #Filtramos en el dataframe el genero que fue solicitado
    if generos.empty:  #Con esta linea nos aseguramos que si para ese genero no hay resultado se notifique
        return f"No se encontraron datos para el género {genero}"
    año_max = generos.loc[generos["playtime_hour"].idxmax()] #Primero identificamos la fila (indice) que tiene la máxima cantidad de horas jugadas para el género dado y posteriormente se selecciona esa fila a partir del indice
    result = {
        'Genero': genero,
        'Año con Más Horas Jugadas': int(año_max["release_year"]),
        'Total de Horas Jugadas': año_max["playtime_hour"]
    }

    return result

PlayTimeGenre("Action")

## Funcion UserForGenre

df_UserForGenre = pd.read_csv("data2\df_UsersForGenre2_final.csv")
df_UserForGenre.columns

def UserForGenre( genero : str ):
    """
    Funcion que devuelve el usuario que acumula más horas jugadas para el género dado 
    y una lista de la acumulación de horas jugadas por año.
    """
    generos2 = df_UserForGenre[df_UserForGenre["main_genre"]== genero]
    user_max = generos2.loc[generos2["playtime_hour"].idxmax()]["user_id"]
    horas_x_año = generos2.groupby(["release_year"])["playtime_hour"].sum().reset_index()
    horas_lista = horas_x_año.rename(columns={"release_year": "Año", "playtime_hour": "Horas"}).to_dict(orient="records")    
    result2 = {
        "Genero": genero,
        "Usuario con Más Horas Jugadas": user_max,
        "Total de Horas Jugadas Por Año": horas_lista
    }
    return result2

UserForGenre("RPG")

df_UserForGenre3 = pd.read_csv("data2/df_UsersForGenre3_final.csv")
df_UserForGenre3.columns

def UserForGenre3( genero : str ):
    """
    Funcion que devuelve el usuario que acumula más horas jugadas para el género dado 
    y una lista de la acumulación de horas jugadas por año.
    """
    generos3 = df_UserForGenre3[df_UserForGenre3["main_genre"]== genero]
    user_max3 = generos3.loc[generos3["playtime_hour"].idxmax()]["user_id"]
    horas_x_año3 = generos3.groupby(["year_posted"])["playtime_hour"].sum().reset_index()
    horas_lista3 = horas_x_año3.to_dict(orient="records")
    result6 = {
        "Genero": genero,
        "Usuario con Más Horas Jugadas": user_max3,
        "Total de Horas Jugadas Por Año": horas_lista3
    }
    return result6

UserForGenre3("Adventure")

## Funcion UsersRecommend

df_UsersRecommend = pd.read_csv("data2/df_UsersRecommend2_final.csv")
df_UsersRecommend

def UsersRecommend( año : int ):
    """
    Funcion que devuelve el top 3 de juegos MÁS recomendados por usuarios para el año dado.
    """
    df_año= df_UsersRecommend[df_UsersRecommend["year_posted"]== año]
    if type(año) != int:
        return {"Debes colocar el año en entero, Ejemplo:2012"}
    if año < df_UsersRecommend["year_posted"].min() or año > df_UsersRecommend["year_posted"].max():
        return {"Año no encontrado"}
    df_ordenado_recomendacion = df_año.sort_values(by="recommendation_count", ascending=False)
    top_3_juegos = df_ordenado_recomendacion.head(3)[["app_name","recommendation_count"]]
    result3 ={
        "Año": año,
        "Top 3 Juegos Más Recomendados": top_3_juegos.to_dict(orient="records")
    }
    return result3

UsersRecommend(2015)

## Funcion UsersWorstDeveloper

df_UsersWorstDeveloper = pd.read_csv("data2/df_UsersWorstDeveloper_final1.csv")
df_UsersWorstDeveloper

def UsersWorstDeveloper( año : int ):
    """
    Funcion que devuelve el top 3 de desarrolladoras con juegos MENOS 
    recomendados por usuarios para el año dado.
    """
    df_año2 = df_UsersWorstDeveloper[df_UsersWorstDeveloper["year_posted"]== año]
    if type(año) != int:
        return {"Debes colocar el año en entero, Ejemplo:2012"}
    if año < df_UsersRecommend["year_posted"].min() or año > df_UsersRecommend["year_posted"].max():
        return {"Año no encontrado "}
    df_ordenado_recomendacion2 = df_año2.sort_values(by="recommendation_count", ascending=False)
    top_3_developers = df_ordenado_recomendacion2.head(3)[["developer","recommendation_count"]]
    result4 = {
        'Año': año,
        'Top 3 Desarrolladoras Menos Recomendadas': top_3_developers.to_dict(orient="records")
    }
    return result4

UsersWorstDeveloper(2015)

## Funciones sentiment_analysis

df_Sentiment_Analysis = pd.read_csv("data2/df_Sentiment_analysis_final.csv")
df_Sentiment_Analysis

def sentiment_analysis( desarrolladora : str ):
    """
    Funcion que devuelve un diccionario con el nombre de la desarrolladora como llave y una lista 
    con la cantidad total de registros de reseñas de usuarios que se encuentren categorizados con 
    un análisis de sentimiento como valor.
    """
    if type(desarrolladora) != str:
        return "Debes colocar un developer de tipo str, EJ:'07th Expansion'"
    if len(desarrolladora) == 0:
        return "Debes colocar un developer en tipo String"
    df_developer = df_Sentiment_Analysis[df_Sentiment_Analysis["developer"]== desarrolladora]
    sentiment_counts = df_developer.groupby("sentiment_analysis")["sentiment_analysis_count"].sum().to_dict()
    sentiment_dicc = {0: "Negativo", 1: "Neutral", 2: "Positivo"}
    sentiment_counts = {sentiment_dicc[key]: value for key, value in sentiment_counts.items()}
    result50 = {desarrolladora: sentiment_counts}
    return result50

sentiment_analysis("Valve")
