==> 05_Modelo_Recomendacion.ipynb <==
# Modelo de Aprendizaje Automático:
Después de completar el Análisis Exploratorio de Datos (EDA) para comprender la naturaleza de nuestros datos, avanzamos hacia la creación de un modelo de recomendación basado en la similitud del coseno. El enfoque del modelo se fundamenta en la similitud entre juegos. Al proporcionar el nombre o identificador de un juego específico, el sistema generará una lista de 5 juegos recomendados que comparten similitudes con el juego de referencia. Esta metodología es ideal para usuarios que buscan descubrir títulos relacionados o similares a sus juegos preferidos.

import pandas as pd
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity
import scipy.sparse as sp
import pyarrow.parquet as pq
import matplotlib.pyplot as plt
import seaborn as sns

## Carga de Datos


df_output = pd.read_csv("data2/output_steam_games_limpio9.csv")
df_output

## Selección de Columnas Relevantes
En esta sección, se seleccionaron las columnas relevantes ("app_name", "id", y "main_genre") del DataFrame original df_output.

df_recomendacion =df_output[["app_name","id","main_genre"]]
df_recomendacion

## Codificación One-Hot
Se aplicó la codificación one-hot a la columna "main_genre" utilizando la función pd.get_dummies.

modelo_recomendacion = pd.get_dummies(df_recomendacion, columns=["main_genre"])
modelo_recomendacion 

## Transformación de Valores Booleanos
Se reemplazaron los valores booleanos por 1 y 0 en el DataFrame modelo_recomendacion.

modelo_recomendacion = modelo_recomendacion.replace({True: 1, False: 0})
modelo_recomendacion

## Modificación de Nombres de Columnas
Se realizaron modificaciones en los nombres de las columnas para quitar el prefijo "main_genre_".

modelo_recomendacion.columns = modelo_recomendacion.columns.str.replace("main_genre_", "")
modelo_recomendacion

modelo_recomendacion.columns

## Función de Recomendación de Juegos
En esta sección se definió una función llamada recomendacion_juego que tiene como objetivo proporcionar recomendaciones de juegos similares en función del ID del juego proporcionado.
* Entrada:
id_juego: El ID del juego para el cual se desea obtener recomendaciones.
* Proceso:
La función verifica si el ID del juego es válido y existe en el DataFrame modelo_recomendacion.
Se realiza un cálculo de similitud coseno entre el juego dado y una muestra aleatoria de juegos.
Se ordenan los juegos en la muestra por similitud y se seleccionan los 5 más similares.
* Salida:
La función devuelve un diccionario con los nombres de los juegos recomendados. Si hay un error, devuelve un mensaje de error correspondiente.

def recomendacion_juego(id_juego):
    try:
        id_juego = int(id_juego)
        juego_seleccionado = modelo_recomendacion[modelo_recomendacion["id"] == id_juego]
        if juego_seleccionado.empty:
            return {"error": f"El juego con el ID '{id_juego}' no se encuentra."}
        indice_juego = juego_seleccionado.index[0]
        muestra = 3000
        df_muestra = modelo_recomendacion.sample(n=muestra, random_state=50)
        juego_features = modelo_recomendacion.iloc[indice_juego, 3:]
        muestra_features = df_muestra.iloc[:, 3:]
        similitud = cosine_similarity([juego_features], muestra_features)[0]
        recomendaciones = sorted(enumerate(similitud), key=lambda x: x[1], reverse=True)[:5]
        recomendaciones_indices = [i[0] for i in recomendaciones]
        recomendaciones_names = df_muestra["app_name"].iloc[recomendaciones_indices].tolist()

        return {"Juegos_similares": recomendaciones_names}

    except ValueError:
        return {"error": "El ID del juego debe ser un número entero válido."}


recomendacion_juego(1250)

# modelo_recomendacion= modelo_recomendacion.to_csv("Datasets/modelo_reco_final.csv")
